Don't depend on jenkins_api_client for job creation
The jenkins_api_client has a dependency list thats a bit steeper than we
noticed, and this isn't going to fly. With this commit we remove our reliance
on the jenkins_api_client for creating a jenkins job. Instead we opt to do it
the old fashioned way using curl. We go back to generating an xml file for the
job configuration, and then curling that up to jenkins to create the job.
Functionally, this is identical to the previous work flow. We also replace the
job existence check with a curl-based check. Because "sh" will raise an
exception if the executed command returns non-zero, we switch curl_form_args to
%x, which will still populate $?.success but won't raise a failure exception,
%since our success case is actually the non-zero return if we are checking for
%a job already existing.