Remove subdir (products|devel etc) from repo config name
The rpm repo config file name is currently
project-version-dist-arch-subdir.repo, where subdir is the subdir in
yum.puppetlabs.com that the package is in, e.g. devel or products. The subdir
was only added for a little extra metadata, but is actually unnecessary. It
also introduces an additional bit of complexity for jobs that are going to use
the configs, because the job has to know if the package built was an RC or a
final, which means parsing the git-describe. This is even more meaningless for
builds on commit because then the 'rc' or 'final' status is just a check of
which is closer - a 'final' tag or an 'rc' tag. This commit removes the subdir
from the repo file name, so now it will be project-version-dist-arch.repo,
making it easier to consume the repo file.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>