Add erb methods, get_binding to Pkg::Config
Currently all the old @build instance variables are set as essentially global
rake application instance variables in order to allow our erb templates to
reference them without qualification. This is dumb. In the new world, we'll
take an optional binding when ERBing, as added in the commit. This commit also
adds a new Pkg::Config.get_binding method to get the class variables of
Pkg::Config, which is what we'll pass to the templates.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>