deprecations, defaults, and reassignments
Add mechanisms by which to support the expression of deprecations, default
values, and reassignments as part of the config. This mostly involves adding
additional constants to the params.rb file, and new methods to handle them in
config.rb. This is largly to replace the requirement for build.rake and
10_setupvar.rake.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>