update 20_setupextravars.rake for new librarification
We've moved the variable reassignment of pe_name => project into the
reassignments in the library. We don't need to export all of our instance
variables anymore, because we're using bindings wherever we need them, and
finally we don't have to haphazardly use a bunch of manual environment variable
overrides - we have a method for that now.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>