Move project root setting into defaults, and don't allow its override
since it's a default, it belongs in the load_defaults method. Also, it is
absolutely essential that the overrides do not override the project_root
setting, because this is environment-specific, and any value in a params file
is going to be wrong. Thus, if we have a project root before we begin
overriding, we save it and restore it after overrides.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>