Fixups in tar.rake
First, if templates is truly an array, doing Dir[<elem>] on each template
causes it to turn into an array of arrays. We just wanted an array of strings,
and I imagine we wanted a fully qualified path to the file. We take care of
doing a qualified path of the element in the tar.rb lib. Finally, the excludes
aren't handling an array like they should be, and what if I want to include
excludes as an array?

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>